improve damage logic material take damage with ANY force objects - and self!

+++ Assets/_Game/_Scripts/Refactored/Base Logic/IDamageable.cs
@@ -2,6 +2,6 @@ using System;
 
 public interface IDamageable
 {
-    public event Action<int> OnTryTakeDamage;  
-    public void TryTakeDamage(int damage);
+    public event Action<float> OnTryTakeDamage;  
+    public void TryTakeDamage(float damage);
 }
